Вынес код для работы с каналом уведомлений в отдельный класс
