+++ lib/data/services/notifications_service.dart
@@ -1,37 +1,16 @@
-import 'package:flutter/services.dart';
+import 'package:todos/data/channels/notifications_channel.dart';
 import 'package:todos/domain/models/todo.dart';
 import 'package:todos/domain/services/i_notifications_service.dart';
 
-/// Класс для вызова нативного кода по работе с уведомлениями.
-class NotificationsService extends INotificationsService {
-  static const platform = MethodChannel('todos/notifications');
-
-  /// Вызывает нативный код для планирования уведомления.
-  ///
-  /// Вызывает метод "scheduleNotification".
-  ///
-  /// Передает id задачи "todo_id", название задачи "title" и время
-  /// уведомления в миллисекундах "time_millis".
+/// Сервис для работы с уведомлениями.
+class NotificationsService implements INotificationsService {
   @override
   Future<bool> scheduleNotification(Todo todo) async {
-    return await platform.invokeMethod('scheduleNotification', {
-      "todo_id": todo.id,
-      "title": todo.title,
-      "time_millis": todo.notificationTime.millisecondsSinceEpoch,
-    });
+    return NotificationsChannel.scheduleNotification(todo);
   }
 
-  /// Вызывает нативный код для отмены уведомления.
-  ///
-  /// Вызывает метод "cancelNotification".
-  ///
-  /// Передает id задачи "todo_id", название задачи "title" и время
-  /// уведомления в миллисекундах "time_millis".  @override
+  @override
   Future<bool> cancelNotification(Todo todo) async {
-    return await platform.invokeMethod('cancelNotification', {
-      "todo_id": todo.id,
-      "title": todo.title,
-      "time_millis": todo.notificationTime.millisecondsSinceEpoch,
-    });
+    return NotificationsChannel.cancelNotification(todo);
   }
 }
